Eliminate need for boilerplate modules.
